refactor(backend): extract status/topic/category crud
Continues the incremental split of the crud monolith. Status, Topic and
Category each move to their own module; callers switch to direct
submodule imports. Pure move -- no logic changes.
